.github/scripts/requirements.txt
frozendict~=2.4.6
Jinja2~=3.1.5
pillow~=11.1.0
